изменений функциональности нет, немного структурировала
