+++ app/handlers/tasks.go
@@ -1,22 +1,9 @@
 package handlers
 
 import (
-	"encoding/json"
 	"github.com/Bostanova/go_final_project/app/repo"
-	"log"
 	"net/http"
 )
-
-func createBody(body map[string]interface{}, w http.ResponseWriter) {
-	bodyJson, err := json.Marshal(body)
-	if err != nil {
-		log.Println(err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	w.WriteHeader(http.StatusOK)
-	w.Write(bodyJson)
-}
 
 func GetTasksHandler(w http.ResponseWriter, r *http.Request) {
 	body := make(map[string]interface{})
@@ -24,16 +11,16 @@ func GetTasksHandler(w http.ResponseWriter, r *http.Request) {
 	tasks, err := repo.DB.GetTasks()
 	if err != nil {
 		body["error"] = err.Error()
-		createBody(body, w)
+		createResponse(body, http.StatusInternalServerError, w)
 		return
 	}
 
 	if tasks == nil {
 		body["tasks"] = []string{}
-		createBody(body, w)
+		createResponse(body, http.StatusOK, w)
 		return
 	}
 
 	body["tasks"] = tasks
-	createBody(body, w)
+	createResponse(body, http.StatusOK, w)
 }
